requirements.txt
fastapi==0.115.3
mavsdk==2.8.4
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
uvicorn==0.32.0
python-dotenv==1.0.1